requirements.txt
Flask
openai
python-dotenv
weasyprint
docxtpl
python-docx
Pillow
